+++ foundation_stdlib/lib/value.lua
@@ -10,11 +10,13 @@ function foundation.com.is_blank(value)
   end
 end
 
+local is_blank = foundation.com.is_blank
+
 --
 -- Takes a list of arguments, and returns the first non-blank one
 function foundation.com.first_present(...)
   for _, value in ipairs({...}) do
-    if not foundation.com.is_blank(value) then
+    if not is_blank(value) then
       return value
     end
   end
@@ -24,7 +26,7 @@ end
 --
 --
 -- @spec deep_equals(Value, Value) :: boolean
-function foundation.com.deep_equals(a, b, depth)
+local function deep_equals(a, b, depth)
   depth = depth or 0
   if depth > 20 then
     error("deep_equals depth exceeded")
@@ -41,7 +43,7 @@ function foundation.com.deep_equals(a, b, depth)
       end
 
       for k, _ in pairs(keys) do
-        if not foundation.com.deep_equals(a[k], b[k], depth + 1) then
+        if not deep_equals(a[k], b[k], depth + 1) then
           return false
         end
       end
@@ -53,3 +55,5 @@ function foundation.com.deep_equals(a, b, depth)
     return false
   end
 end
+
+foundation.com.deep_equals = deep_equals